Fix a nasty bug in CameraData serialization/deserialization.
The 'size_t' data type has a different size on my laptop (8 bits)
than it does on the Raspberry Pi 2 (4 bits). This caused the
the spacing of data in the zmq message to be different on my laptop
than on the Pi, leading to deserialization failure. The fix was
to switch all datatypes that get serialized to types of explicit size,
e.g. uint8_t and uint16_t.
Also added an extra form of data validation in these zmq messages -
each message type has its own "validation code", which provides
immediate verification that the received message is indeed a
serialized vector of CameraData.