Make cameraServer take parameters from config class
Took a new approach to the config class, centered on macros that eliminate a significant amount of boilerplate code.

Migrated AdaVision config to this new global config format and restructured the config file format significantly.

Expanded the unit tests for the config file, to ensure that both the master example config file and the production config file are checked for consistency if any further changes to the config file format are made.